Reuse menu to blocks conversion util on navigation-link block
